Move items out of compat: url_parse/url_unquote and more
