+++ tests/test_in_wsgiref.py
@@ -2,10 +2,11 @@ import cgi
 import logging
 import socket
 import sys
+from urllib.request import urlopen as url_open
 
 import pytest
 
-from webob.compat import Empty, Queue, url_open
+from webob.compat import Empty, Queue
 from webob.request import Request
 from webob.response import Response
 from webob.util import bytes_
